# lib/swagger-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Prueba Tecnica Fullstack API",
    "version": "1.0.0",
    "description": "API documentation for the fullstack application"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://prueba-fs.vercel.app",
      "description": "Production server"
    }
  ],
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "User ID"
          },
          "name": {
            "type": "string",
            "description": "User name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "User email"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "admin"
            ],
            "description": "User role"
          }
        }
      },
      "Transaction": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Transaction ID"
          },
          "concepto": {
            "type": "string",
            "description": "Transaction concept"
          },
          "monto": {
            "type": "number",
            "description": "Transaction amount"
          },
          "fecha": {
            "type": "string",
            "format": "date-time",
            "description": "Transaction date"
          },
          "userId": {
            "type": "string",
            "description": "User ID who created the transaction"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/api/profile": {
      "get": {
        "summary": "Obtener perfil del usuario actual",
        "description": "Recuperar la información del perfil del usuario actual",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Perfil de usuario con estadísticas",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/User"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "statistics": {
                          "type": "object",
                          "properties": {
                            "transactionCount": {
                              "type": "integer",
                              "description": "Number of transactions"
                            },
                            "totalAmount": {
                              "type": "number",
                              "description": "Total amount of transactions"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      },
      "put": {
        "summary": "Actualizar perfil del usuario actual",
        "description": "Actualizar la información del perfil del usuario actual",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "email"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "User name"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "User email"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Perfil actualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Faltan campos requeridos"
          },
          "401": {
            "description": "No autorizado"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
    "/api/transactions": {
      "get": {
        "summary": "Obtener todas las transacciones",
        "description": "Recuperar una lista de todos los registros de transacciones",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de transacciones",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Transaction"
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      },
      "post": {
        "summary": "Crear una nueva transacción",
        "description": "Crear un nuevo registro de transacción (solo administrador)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "concepto",
                  "monto",
                  "fecha"
                ],
                "properties": {
                  "concepto": {
                    "type": "string",
                    "description": "Transaction concept"
                  },
                  "monto": {
                    "type": "number",
                    "description": "Transaction amount"
                  },
                  "fecha": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Transaction date"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Transacción creada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          },
          "400": {
            "description": "Faltan campos requeridos"
          },
          "401": {
            "description": "No autorizado"
          },
          "403": {
            "description": "Prohibido - Se requiere acceso de administrador"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
    "/api/users": {
      "get": {
        "summary": "Obtener todos los usuarios",
        "description": "Recuperar una lista de todos los usuarios (solo administrador)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de usuarios",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          },
          "403": {
            "description": "Prohibido - Se requiere acceso de administrador"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      },
      "put": {
        "summary": "Actualizar un usuario",
        "description": "Actualizar información del usuario (solo administrador)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "id",
                  "name",
                  "role"
                ],
                "properties": {
                  "id": {
                    "type": "string",
                    "description": "User ID"
                  },
                  "name": {
                    "type": "string",
                    "description": "User name"
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "user",
                      "admin"
                    ],
                    "description": "User role"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuario actualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Faltan campos requeridos"
          },
          "401": {
            "description": "No autorizado"
          },
          "403": {
            "description": "Prohibido - Se requiere acceso de administrador"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    }
  },
  "tags": []
}